Simplify the showHomepage test
Inside of TestShowHomepage, there is arbitrary code that creates a
request to http://localhost:8080. However, this leads to the impression
that that value is actually necessary in order for the test to pass,
which it isn’t. Instead, pass in an empty string and change the comment
to something a little more descriptive.