Add a separate GitHub-oriented start file and more
Create githubstart.go, an almost-identical remake of start.go, except
with a different import location. This file is intended to make it
easier to clone Sonobabble from GitHub, without having to worry about
moving it to a different directory.

Also, try (again, for the 5th time) to fix findAbsolutePath, this time
by adding more detailed errors.