Create a contributing guide
Create a guide for contributing to Sonobabble’s development. Although
Sonobabble is still in prealpha, this guide can help those who want to
help or are just simply interested. GitHub will automatically show the
guide when a viewer goes to create a pull request or issue.

The guide will cover commit messages, pull requests, and code styling
such as line-length and error messages.