Add the Sonobabble package name to errors
Instead of creating an error in the form

	fmt.Errorf("generateRandomGoPath: %s", pathError)

create it in the form

	fmt.Errorf("sonobabble.generateRandomGoPath: %s", pathError)

(with “sonobabble” inserted at the front).

This makes some error messages to more closely align with others from
functions in other packages, which often list the original package in a
way like

	fmt.Errorf("os.Stat %s: %s", location, statError)

This also makes it easier to see where a function that generated an
error came from.